Make idman also associate all the commits
Assocman now doesn't take a separate file of identity results anymore,
instead it uses the first line of its input for that. Analogously,
graphman now outputs its results on a single line of JSON and assocman
has a pretty multi-line output.

The idman script has been completely rewritten to handle all the
time-travel-laden pipery between parseman, graphman and assocman, but
the usage is still the same.