Rewrite graphman to be less graphy
Doesn't use an actual graph anymore, instead it just iterates over the
set of identities and merges them until there's none left to merge.
While it's not as pretty on an abstract level ot not have an actual
graph, it makes the algorithm implementation much simpler, especially
if they need to process the identities.