Use [name, e-mail address] tuples
Since we figured out that nobody ever signs a commit ever, we can toss
the handling of signer and signer keys. That means all that's left is
the name and e-mail address of the author and committer.

So instead of combining all artifacts in one big pot, idman and assocman
now make the distinction between names and e-mails. They handle them as
tuples (that is, arrays with two elements) all the way through.

This makes graphman even more general and probably lets us implement
even more algorithms, since many of them do in fact use these kinds of
tuples.

The identity merging algorithm can also be as bad as it wants to be now,
since a badly merged input doesn't cause assocman to get confused by
ambiguitiy anymore.